[Improvement] Remove start folder
Automatic cd to init folder is not longer necessary. This is causing
problems when opening shell in an specific folder. For the future I
could check the current folder and if different from the default then it
might update the init folder. For the moment this is not required
anymore.